requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
pyspark==3.5.0